Done with Binary Tree with Java
removed the get max from the class, something to implement in an algorithm.  Considering it's generic, how to find the max of BinaryTree<Person> is something the binary tree shouldn't have to do, but the user.  This just holds the data, more complex algorithms (with generic data types will require the user to do)